Refactor: Convert `orderDate` to LocalDate, adjust date search, and add JWT authentication
- Changed the `orderDate` field in the `Order` entity to use `LocalDate` instead of `Date`, ensuring only the date is stored without the time.
- Implemented conversion of `String` to `LocalDate` in the controller and DTO when receiving the input date.
- Adjusted the repository query to correctly handle dates without including the time.
- Fixed the behavior of order creation to accept the date as a parameter, ensuring the creation date is correctly recorded.
- Modified the date search endpoint to only consider the order date without affecting the creation time.
- Added a JWT authentication filter (`JwtAuthenticationFilter`) to validate tokens in incoming requests by extracting and verifying the JWT from the authorization header.

This change improves consistency in date handling, adds security through JWT authentication, and enables precise date queries without considering the time.

+++ backend/order_service/src/main/java/ucv/app_inventory/order_service/application/OrderDeleteUseCase.java
@@ -3,6 +3,7 @@ package ucv.app_inventory.order_service.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ucv.app_inventory.order_service.exception.OrderNotFoundException;
 import ucv.app_inventory.order_service.infrastructure.outbound.database.OrderMySqlRepository;
 
 @Service
@@ -12,7 +13,10 @@ public class OrderDeleteUseCase {
     private final OrderMySqlRepository orderMySqlRepository;
 
     @Transactional
-    public void eliminarPedido(Long id) {
-        orderMySqlRepository.deleteById(id);
+    public void deleteOrder(Long id) {
+        if (!orderMySqlRepository.existsById(id)) {
+            throw new OrderNotFoundException("Order with ID " + id + " not found");
+        }
+        orderMySqlRepository.deleteById(id); // Esto debería eliminar también los detalles si tienes cascade delete configurado
     }
 }
